requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.31.0
sqlalchemy>=2.0.36
aiosqlite>=0.20.0
asyncpg>=0.30.0
jinja2>=3.1.4
python-multipart>=0.0.12
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
bcrypt>=4.2.0
aiogram>=3.17.0
python-dotenv>=1.0.0
